Detect pcre using pkg-config
The current approach uses extra-libraries and a configure script, which
makes this package difficult to install on Windows since you need to
specify where libpcre lives manually using the --extra-lib-dirs flag.

Instead, we can use pkg-config, which automatically configures the right
linker and include flags and completely eliminates the need for a
configure script. As a result, this fixes #2, since that issue becomes
moot.